+++ src/routes/+page.svelte
@@ -4,7 +4,7 @@
 	import Library_Footer from '$lib/Library_Footer.svelte';
 	import Card from '$lib/Card.svelte';
 	import {pkg_context} from '$routes/pkg.js';
-	import Svg from '$lib/Svg.svelte';
+	import Svg, {type Svg_Data} from '$lib/Svg.svelte';
 	import {
 		fuz_logo,
 		fuz_code_logo,
@@ -12,11 +12,31 @@
 		fuz_mastodon_logo,
 		fuz_gitops_logo,
 		fuz_template_logo,
+		gro_logo,
 		moss_logo,
 	} from '$lib/logos.js';
 	import Hidden_Personal_Links from '$lib/Hidden_Personal_Links.svelte';
 
 	const pkg = pkg_context.get();
+
+	interface Project_Item {
+		name: string;
+		url: string;
+		logo: Svg_Data;
+		description: string;
+		glyph: string;
+	}
+
+	// prettier-ignore
+	const project_items: Array<Project_Item> = [ 
+		{name: 'moss', url: 'https://moss.ryanatkn.com/', logo: moss_logo, description: 'CSS framework and design system', glyph: '🌿'},
+		{name: 'gro', url: 'https://gro.ryanatkn.com/', logo: gro_logo, description: 'task runner and toolkit extending SvelteKit', glyph: '🌰'},
+		{name: 'fuz_template', url: 'https://template.fuz.dev/', logo: fuz_template_logo, description: 'a static web app and Node library template with TypeScript, Svelte, SvelteKit, Vite, esbuild, Fuz, and Gro', glyph: '❄'},
+		{name: 'fuz_blog', url: 'https://blog.fuz.dev/', logo: fuz_blog_logo, description: 'blog software from scratch with SvelteKit', glyph: '🖊️'},
+		{name: 'fuz_mastodon', url: 'https://mastodon.fuz.dev/', logo: fuz_mastodon_logo, description: 'Mastodon components and helpers for Svelte, SvelteKit, and Fuz', glyph: '🦣'},
+		{name: 'fuz_code', url: 'https://code.fuz.dev/', logo: fuz_code_logo, description: 'syntax styling utilities and components for TypeScript, Svelte, SvelteKit', glyph: '🎨'},
+		{name: 'fuz_gitops', url: 'https://gitops.fuz.dev/', logo: fuz_gitops_logo, description: 'a tool for managing many repos', glyph: '🪄'},
+	];
 </script>
 
 <main class="box w_100">
@@ -28,10 +48,10 @@
 			<div class="mb_lg p_xs2 shadow_md radius_sm">
 				<div class="width_sm panel p_lg shadow_inset_xs bg">
 					<p>
-						Fuz is a Svelte UI library with components and helpers for making friendly and helpful
-						websites. It builds on my CSS framework <a href="https://moss.ryanatkn.com/">Moss</a>,
-						which is based on CSS custom properties, and Fuz continues those patterns. It's in early
-						alpha and will have many breaking changes.
+						Fuz is a Svelte UI library with components and helpers for making modern websites. It
+						builds on my CSS framework <a href="https://moss.ryanatkn.com/">Moss</a>, which is based
+						on CSS custom properties, and Fuz continues those patterns. It's in early alpha and will
+						have many breaking changes.
 					</p>
 					<p>
 						Fuz is being made to support <a href="https://www.zzzbot.dev/">Zzz</a> and
@@ -56,30 +76,9 @@
 			<!-- TODO ideally this wouldn't duplicate metadata like descriptions, but adding fuz_gitops to this repo is heavy -->
 			<h2 class="mt_0 mb_xl2 px_md">other packages</h2>
 			<menu class="unstyled size_lg">
-				<a class="row chip bg px_md py_xs radius_sm mb_lg" href="https://moss.ryanatkn.com/"
-					><Svg data={moss_logo} size="var(--icon_size_lg)" />
-					<span class="ml_lg">moss</span></a
-				>
-				<a class="row chip bg px_md py_xs radius_sm mb_lg" href="https://template.fuz.dev/"
-					><Svg data={fuz_template_logo} size="var(--icon_size_lg)" />
-					<span class="ml_lg">fuz_template</span></a
-				>
-				<a class="row chip bg px_md py_xs radius_sm mb_lg" href="https://blog.fuz.dev/"
-					><Svg data={fuz_blog_logo} size="var(--icon_size_lg)" />
-					<span class="ml_lg">fuz_blog</span></a
-				>
-				<a class="row chip bg px_md py_xs radius_sm mb_lg" href="https://mastodon.fuz.dev/"
-					><Svg data={fuz_mastodon_logo} size="var(--icon_size_lg)" />
-					<span class="ml_lg">fuz_mastodon</span></a
-				>
-				<a class="row chip bg px_md py_xs radius_sm mb_lg" href="https://code.fuz.dev/"
-					><Svg data={fuz_code_logo} size="var(--icon_size_lg)" />
-					<span class="ml_lg">fuz_code</span></a
-				>
-				<a class="row chip bg px_md py_xs radius_sm" href="https://gitops.fuz.dev/"
-					><Svg data={fuz_gitops_logo} size="var(--icon_size_lg)" />
-					<span class="ml_lg">fuz_gitops</span></a
-				>
+				{#each project_items as project_item}
+					{@render package_thumbnail(project_item)}
+				{/each}
 			</menu>
 		</section>
 		<section>
@@ -93,16 +92,38 @@
 	</div>
 </main>
 
+{#snippet package_thumbnail(project_item: Project_Item)}
+	<a class="thumbnail row bg px_md py_xs radius_sm mb_lg" href={project_item.url}
+		><Svg shrink={false} data={project_item.logo} size="var(--icon_size_lg)" />
+		<div class="pl_lg width_sm">
+			<div class="thumbnail_name">{project_item.name}</div>
+			<div class="thumbnail_description size_md text_color_3 font_weight_500">
+				{project_item.description}
+				{project_item.glyph}
+			</div>
+		</div></a
+	>
+{/snippet}
+
 <style>
-	a.chip {
+	.thumbnail {
 		box-shadow: var(--shadow_bottom_xs)
 			color-mix(in hsl, var(--shadow_color) var(--shadow_alpha_1), transparent);
 	}
-	a.chip:hover {
+	.thumbnail:hover {
 		box-shadow: var(--shadow_bottom_sm)
 			color-mix(in hsl, var(--shadow_color) var(--shadow_alpha_1), transparent);
+
+		/* show the underline only on the name, not the description */
+		text-decoration: none;
+		.thumbnail_name {
+			text-decoration: underline;
+		}
+		.thumbnail_description {
+			text-decoration: none;
+		}
 	}
-	a.chip:active {
+	.thumbnail:active {
 		box-shadow: var(--shadow_top_xs)
 			color-mix(in hsl, var(--shadow_color) var(--shadow_alpha_1), transparent);
 	}