+++ src/routes/library/tomes.ts
@@ -19,7 +19,9 @@ import logos from '$routes/library/logos/+page.svelte';
 
 // TODO maybe decouple `related` from `Tome` to get bidirectionality for free
 
-export const tomes: Tome[] = [
+// TODO @many add `Svg` library entry
+
+export const tomes: Array<Tome> = [
 	{
 		name: 'introduction',
 		category: 'guide',